Update: User alter tg_id field; update: TelegramBot auth_complete with django orm

+++ telegramservice/users/models.py
@@ -20,7 +20,7 @@ class User(AbstractUser):
         expired = "expired", _("Expired")
 
     # fields
-    tg_id = models.IntegerField(
+    tg_id = models.BigIntegerField(
         _("Telegram ID"), db_index=True, unique=True, null=True, blank=True
     )
     name = models.CharField(_("Name of User"), blank=True, max_length=255)
